Fix float16 pixel_access::get()

+++ include/visionaray/detail/pixel_access.h
@@ -1015,7 +1015,7 @@ inline void get(
 }
 
 //-------------------------------------------------------------------------------------------------
-// Get AVX simd vector from scalar buffer
+// Get AVX-512 simd vector from scalar buffer
 // TODO: consolidate w/ float4 version
 //
 
@@ -1032,8 +1032,8 @@ inline void get(
         T const*                        buffer
         )
 {
-    const int w = packet_size<simd::float8>::w;
-    const int h = packet_size<simd::float8>::h;
+    const int w = packet_size<simd::float16>::w;
+    const int h = packet_size<simd::float16>::h;
 
     simd::aligned_array_t<simd::float16> out;
